Refactor to avoid erroring during tests
